refactor: rename DoneUntil to Done and update semantics
- Rename SetDoneUntil/GetDoneUntil/WaitForDoneUntil to SetDone/GetDone/WaitForDone
- Change semantics: Done N means cycle N is complete (was: N-1 complete)
- Change initial value from 0 to -1
- Change all uint64 to int for cycle-related types
- Fix bug in multi_upstream_port.go: use pkt.Cycle-1 instead of pkt.Cycle when draining packets in notification handler

+++ internal/dataflow/packet/packet.go
@@ -15,6 +15,6 @@ type Packet struct {
 // PacketWithCycle associates a packet with the cycle in which it becomes visible to
 // the destination flow. This is the element transmitted through link channels.
 type PacketWithCycle struct {
-	Cycle  uint64
+	Cycle  int
 	Packet Packet
 }
